// components/admin/adminMainCategoryList/adminMainCategoryListStyle.scss.d.ts
export type Styles = {
  'groupContainer': string;
  'groupContainerItems': string;
  'groupTitle': string;
  'imageAttestationNewEmployees': string;
  'imageBeginChatColor': string;
  'imageBookshelves': string;
  'imageBuildingBlocks': string;
  'imageCertificate': string;
  'imageContainer': string;
  'imageEducator': string;
  'imageNotebook': string;
  'imagePeoples': string;
  'imageProject': string;
  'imageProjectRegistry': string;
  'imageQuestions': string;
  'imageResourceAlignment': string;
  'imageTeamwork': string;
  'imageUndrawScrumBoard': string;
  'imageUnknown': string;
  'mainCategoryItem': string;
  'textContainer': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;